+++ etna/analysis/feature_selection/mrmr_selection.py
@@ -72,6 +72,12 @@ def mrmr(
     relevance_aggregation_fn = AGGREGATION_FN[AggregationMode(relevance_aggregation_mode)]
     redundancy_aggregation_fn = AGGREGATION_FN[AggregationMode(redundancy_aggregation_mode)]
 
+    # can't compute correlation of categorical column with the others
+    try:
+        regressors = regressors.astype(float)
+    except ValueError as e:
+        raise ValueError(f"Only convertible to float features are allowed! Error: {str(e)}")
+
     relevance = relevance_table.apply(relevance_aggregation_fn).fillna(0)
 
     all_features = relevance.index.to_list()
@@ -81,14 +87,6 @@ def mrmr(
 
     redundancy_table = pd.DataFrame(np.inf, index=all_features, columns=all_features)
     top_k = min(top_k, len(all_features))
-
-    # can't compute correlation of categorical column with the others
-    cat_cols = regressors.dtypes[regressors.dtypes == "category"].index
-    for cat_col in cat_cols:
-        try:
-            regressors[cat_col] = regressors[cat_col].astype(float)
-        except ValueError:
-            raise ValueError(f"{cat_col} column cannot be cast to float type! Please, use encoders.")
 
     for i in range(top_k):
         score_numerator = relevance.loc[not_selected_features]
